+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Removing aggregation result print statement
 - Created local.yaml file to setup environment variables when running processes locally
 - Implemented OCLC other editions call
+- Added functionality for locally generating Counter 5 downloads reports to analytics folder
 
 ## 2024-08-06 -- v0.13.1
 ## Added